Increased code reliability for 'finalize_invoice_move_lines' function
Increased code reliability for 'finalize_invoice_move_lines' function. This resolves the following problem:

When comparing float number data is possible that the condition 'a == b' doesn't give the attended result, probably due to the underlaying way the programming language and the OS manage float types.
Actually it happened to us trying to validate an invoice with amoun_tax = 366.15; In that case, the condition 366.15 == 366.15 returned False.